+++ app/react-native/src/bin/storybook-start.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 /* eslint-disable no-console */
-import { exec } from 'child_process';
 import path from 'path';
 import program from 'commander';
 import Server from '../server';
@@ -8,19 +7,11 @@ import Server from '../server';
 program
   .option('-h, --host <host>', 'host to listen on')
   .option('-p, --port <port>', 'port to listen on')
-  .option('--haul <configFile>', 'use haul with config file')
-  .option('--platform <ios|android|all>', 'build platform-specific build')
   .option('-s, --secured', 'whether server is running on https')
   .option('-c, --config-dir [dir-name]', 'storybook config directory')
-  .option('--metro-config [relative-config-path]', 'Metro Bundler Custom config')
   .option('-e, --environment [environment]', 'DEVELOPMENT/PRODUCTION environment for webpack')
-  .option('-r, --reset-cache', 'reset react native packager')
-  .option('--skip-packager', 'run only storybook server')
   .option('-i, --manual-id', 'allow multiple users to work with same storybook')
   .option('--smoke-test', 'Exit after successful start')
-  .option('--packager-port <packagerPort>', 'Custom packager port')
-  .option('--root [root]', 'Add additional root(s) to be used by the packager in this project')
-  .option('--projectRoots [projectRoots]', 'Override the root(s) to be used by the packager')
   .parse(process.argv);
 
 const projectDir = path.resolve();
@@ -48,61 +39,3 @@ server.listen(...listenAddr, err => {
     process.exit(0);
   }
 });
-
-if (!program.skipPackager) {
-  let symlinks = [];
-
-  let roots = [projectDir];
-
-  if (program.root) {
-    roots = roots.concat(program.root.split(',').map(root => path.resolve(root)));
-  }
-
-  try {
-    // eslint-disable-next-line global-require
-    require('babel-register')({
-      presets: [require.resolve('babel-preset-flow')],
-      ignore: false,
-      babelrc: false,
-    });
-
-    // eslint-disable-next-line global-require
-    const findSymlinkedModules = require('react-native/local-cli/util/findSymlinkedModules');
-    symlinks = roots.reduce((arr, rootPath) => arr.concat(findSymlinkedModules(rootPath, roots)), [
-      ...roots,
-    ]);
-  } catch (e) {
-    console.warn(`Unable to load findSymlinksPaths: ${e.message}`, e);
-  }
-
-  let projectRoots = (configDir === projectDir ? [] : [configDir]).concat(symlinks);
-
-  if (program.projectRoots) {
-    projectRoots = projectRoots.concat(
-      program.projectRoots.split(',').map(root => path.resolve(root))
-    );
-  }
-
-  let cliCommand = 'react-native start';
-
-  if (program.metroConfig) {
-    cliCommand += ` --config ${program.metroConfig}`;
-  }
-
-  if (program.haul) {
-    const platform = program.platform || 'all';
-    cliCommand = `haul start --config ${program.haul} --platform ${platform}`;
-  }
-  // RN packager
-  exec(
-    [
-      cliCommand,
-      `--projectRoots ${projectRoots.join(',')}`,
-      program.resetCache && '--reset-cache',
-      program.packagerPort && `--port=${program.packagerPort}`,
-    ]
-      .filter(x => x)
-      .join(' '),
-    { async: true }
-  );
-}